CLN: Fix typo
